packages/requirements.txt
numba==0.48.0
numpy==1.18.4
pandas==1.1.1
scikit-learn==0.23.1
scipy==1.4.1
tqdm~=4.48.2
ipython==7.16.1
librosa==0.8.0
webrtcvad==2.0.10
sox
torch==1.6.0
torchvision==0.7.0
matplotlib==3.1.1
seaborn~=0.10.1
resemblyzer~=0.1.1.dev0
Cython==0.29.21
hdbscan==0.8.26
azure-cognitiveservices-speech==1.13.0
Babel==2.8.0
cached-property==1.5.1
cachetools==4.1.1
attrs==19.3.0
cattrs==0.9.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coverage==5.1
crcmod==1.7
databricks-cli==0.11.0
decorator==4.4.2
distlib==0.3.0
dnspython==1.16.0
flake8==3.8.3
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.13
Flask-WTF==0.14.3
future==0.18.2
google-api-core==1.22.0
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-speech==1.3.2
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==19.10.0
idna==2.10
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.0
jedi==0.17.2
Jinja2==2.10.3
joblib==0.15.1
json-merge-patch==0.2
llvmlite==0.31.0
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
moviepy==1.0.0
pendulum==1.4.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
pexpect==4.8.0
pg8000==1.13.2
Pillow==7.1.2
proglog==0.1.9
SoundFile==0.10.3.post1
SQLAlchemy==1.3.18
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.7
ptyprocess==0.6.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
typing
sounddevice
pysoundfile
uuid==1.30
pydub==0.24.1
pyflakes==2.2.0
pylint==2.5.2
pyrsistent==0.16.0
PyYAML==5.3.1
requests==2.24.0
resampy==0.2.2
rsa==4.6
six==1.15.0
srt==3.0.0
tenacity==4.12.0
threadpoolctl==2.1.0
thrift==0.13.0
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.21
wget==3.2
wrapt==1.12.1
WTForms==2.3.1
yapf==0.30.0
zipp==3.1.0
black
